# ⚡️ Speed up function `cells_to_html` by 8%